moved formatPrice to valueFormat
