+++ src/utils/ADempiere/numberFormat.js
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils.js'
+import { formatPrice, formatQuantity } from '@/utils/ADempiere/valueFormat.js'
 import {
   // currencies
   FIELDS_CURRENCY,
@@ -73,87 +74,12 @@ export function formatNumber({
   return formattedNumber
 }
 
-/**
- * Get formatted price and show currency
- * @param {number} value
- * @param {string} currency default 'USD'
- * @returns {string} number format with thousands separator, precision, currency format
- */
-export function formatPrice(value, currency, country) {
-  if (isEmptyValue(value)) {
-    value = 0
-  }
-
-  if (isEmptyValue(currency)) {
-    currency = getCurrency()
-  }
-
-  // const precision = getStandardPrecision()
-
-  if (isEmptyValue(country)) {
-    country = getCountryCode()
-  }
-
-  // get formatted currency number
-  return new Intl.NumberFormat(country, {
-    style: 'currency',
-    currency,
-    useGrouping: true,
-    // minimumFractionDigits: precision,
-    // maximumFractionDigits: precision,
-    minimumIntegerDigits: 1
-  }).format(value)
-}
-
 export function getTaxAmount(basePrice, taxRate) {
   if (isEmptyValue(basePrice) || isEmptyValue(taxRate)) {
     return 0
   }
 
   return (basePrice * taxRate) / 100
-}
-
-/**
- * Get formatted number, integer and decimal
- * @param {number} value
- * @returns {string} number format with thousands separator, precision
- * @returns {boolean} number format without precision
- */
-export function formatQuantity(value, isInteger = false) {
-  if (isEmptyValue(value)) {
-    value = 0
-  }
-
-  let precision = getStandardPrecision()
-  // without decimals
-  if (isInteger) {
-    // if (Number.isInteger(value)) {
-    precision = 0
-  }
-
-  // get formatted decimal number
-  return new Intl.NumberFormat(undefined, {
-    useGrouping: true, // thousands separator
-    minimumIntegerDigits: 1,
-    minimumFractionDigits: precision,
-    maximumFractionDigits: precision
-  }).format(value)
-}
-
-/**
- * Format percentage based on Intl library
- * @param {number} value
- * @returns {string} number format with percentage
- */
-export function formatPercent(value) {
-  if (isEmptyValue(value)) {
-    value = 0
-  }
-
-  // get formatted number
-  return new Intl.NumberFormat(getCountryCode(), {
-    style: 'percent'
-  }).format(value)
 }
 
 /**
